[update] カメラをコンポーネント化
